Use poetry for project and dependency management, restructure repository
