+++ air-quality-backend/README.md
@@ -1,6 +1,12 @@
 # Air Quality Backend
 
+Please follow [Local Development Setup](#local-development-setup) instructions if this is the first time working on the project.
+
 ## Common Commands
+
+### Update Conda environment
+
+`conda env update --name air-quality-dev --file conda/etl_environment.yml --prune`
 
 ### Run lint check
 `python -m flake8`
@@ -8,42 +14,36 @@
 ### Run code formatter
 `python -m black src tests scripts`
 
-### Run all tests in etl tests folder
-`python -m pytest --cov=src/etl --cov=scripts tests/etl_tests`
-
-### Run all tests in api tests folder
-`python -m pytest --cov=src/api tests/api_tests`
+### Run all tests
+`python -m pytest --cov=src/air_quality --cov=scripts tests`
 
 ### Run Fast API
 Follow the tutorial [here](docs/run_fast_api_tutorial.md)
 
-## Conda
+## Local Development Setup
 
-### Install conda
+### 1. Install Conda
 Follow the [installation instructions](https://docs.anaconda.com/free/miniconda/) for conda.
 
-### To create environment
-From within air-quality-backend
-#### Etl Dev Environment
-`conda env create -f conda/etl_environment.yml -n etl`
-#### API Dev Environment
-`conda env create -f conda/api_environment.yml -n api`
+### 2. Setup Conda Environment
 
-### To update
-#### Etl Dev Environment
-`conda env update --name etl --file conda/etl_environment.yml --prune`
-#### API Dev Environment
-`conda env update --name api --file conda/api_environment.yml --prune`
+#### Create Conda environment
+From within air-quality-backend:
 
-### Point PyCharm to your environment
+`conda env create -f conda/etl_environment.yml -n air-quality-dev`
+
+#### Activate Conda environment - CLI
+`conda activate air-quality-dev`
+
+#### Activate Conda environment - PyCharm
 Follow these [instructions](https://www.jetbrains.com/help/pycharm/conda-support-creating-conda-virtual-environment.html#conda-requirements).
 
+### 3. Install dependencies
+`poetry install`
 
-## Local Environment
+### 4. Setup .env file
 
-### Setup .env file
-
-Create
+#### Create .env file
 
 `New-Item -ItemType file -Name ".env"`
 
@@ -54,13 +54,20 @@ MONGO_DB_NAME=air_quality_dashboard_db
 OPEN_AQ_API_KEY=<api_key>
 ```
 
-### CDS API access (CAMS)
+### 5. Get CDS API access (CAMS)
 Create your .cdsapirc file as detailed [here](https://ads.atmosphere.copernicus.eu/api-how-to).
 
-### Setup auto format on save (PyCharm)
+### 6. Get OpenAQ access
+Register for a key [here](https://api.openaq.org/register).
 
+Once you receive your key, place it in your **.env** file
+
+### 7. Setup auto format on save (optional) 
+
+#### PyCharm
 - Go to File -> Settings -> Tools -> Black
 - Select your conda env as the python interpreter
 - Check 'On save'
 - Press Apply
+
