--- air-quality-backend/conda/api_environment.yml
@@ -1,16 +1,0 @@
-name: api
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - black=24.4.2
-  - coverage=7.5.0
-  - fastapi=0.111.0
-  - flake8=7.0.0
-  - pip=24.0
-  - pymongo=4.6.3
-  - pytest=8.1.1
-  - pytest-cov=5.0.0
-  - pytest-mock=3.14.0
-  - python=3.11.9
-  - python-dotenv=1.0.1
