+++ air-quality-backend/conda/etl_environment.yml
@@ -3,28 +3,9 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - black=24.4.2
   - bottleneck=1.3.8
-  - cerberus=1.3.5
   - cfgrib=0.9.11.0
-  - coverage=7.5.0
   - dask=2024.4.2
-  - flake8=7.0.0
   - netcdf4=1.6.5
-  - pip=24.0
-  - pymongo=4.6.3
-  - pytest=8.1.1
-  - pytest-cov=5.0.0
-  - pytest-mock=3.14.0
-  - python=3.11.9
-  - python-dotenv=1.0.1
-  - requests=2.31.0
-  - scipy=1.13.0
+  - poetry=1.8
   - xarray=2024.3.0
-  - pip:
-      - cads-api-client==1.0.2
-      - cdsapi==0.7.0
-      - freezegun==1.5.1
-      - multiurl==0.3.1
-      - tqdm==4.66.4
-      - typing-extensions==4.11.0
