--- air-quality-backend/conda/etl_poetry_env.yml
@@ -1,12 +1,0 @@
-name: etl
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - bottleneck=1.3.8
-  - cfgrib=0.9.11.0
-  - dask=2024.4.2
-  - netcdf4=1.6.5
-  - poetry=1.8
-  - python=3.11.9
-  - xarray=2024.3.0
