--- air-quality-backend/conda/utility_environment.yml
@@ -1,12 +1,0 @@
-name: utility-env
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - black=24.4.2
-  - coverage=7.5.0
-  - flake8=7.0.0
-  - pytest=8.1.1
-  - pytest-cov=5.0.0
-  - pytest-mock=3.14.0
-  - python=3.11.9
