+++ air-quality-backend/scripts/run_forecast_etl.py
@@ -3,10 +3,10 @@ from dotenv import load_dotenv
 from itertools import chain
 import logging
 from logging import config
-from database.locations import get_locations_by_type
-from database.forecasts import insert_data
-from etl.forecast.forecast_dao import fetch_forecast_data
-from etl.forecast.forecast_adapter import transform
+from air_quality.database.locations import get_locations_by_type
+from air_quality.database.forecasts import insert_data
+from air_quality.etl.forecast.forecast_dao import fetch_forecast_data
+from air_quality.etl.forecast.forecast_adapter import transform
 
 config.fileConfig("./logging.ini")
