+++ air-quality-backend/scripts/run_purge_old_data.py
@@ -2,10 +2,10 @@ from datetime import datetime, timedelta
 from dotenv import load_dotenv
 import logging
 from logging import config
-from database.forecasts import delete_data_before as delete_forecast_data
-from database.in_situ import delete_data_before as delete_in_situ_data
+from air_quality.database.forecasts import delete_data_before as delete_forecast_data
+from air_quality.database.in_situ import delete_data_before as delete_in_situ_data
 
-config.fileConfig("../logging.ini")
+config.fileConfig("./logging.ini")
 
 
 def main():
