+++ air-quality-backend/src/air_quality/database/locations.py
