+++ air-quality-backend/src/air_quality/etl/in_situ/openaq_adapter.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 import logging
-from etl.air_quality_index.calculator import (
+from air_quality.etl.air_quality_index.calculator import (
     get_pollutant_index_level,
     get_overall_aqi_level,
 )
-from etl.air_quality_index.pollutant_type import PollutantType
-from etl.in_situ.sort_in_situ import sort_by_distance_and_time
+from air_quality.etl.air_quality_index.pollutant_type import PollutantType
+from air_quality.etl.in_situ.sort_in_situ import sort_by_distance_and_time
 
 required_pollutant_data = {
     "o3": PollutantType.OZONE,
