+++ air-quality-backend/tests/etl_tests/forecast/forecast_data_test.py
@@ -1,7 +1,7 @@
 import pytest
 import xarray
-from etl.air_quality_index.pollutant_type import PollutantType
-from etl.forecast.forecast_data import (
+from air_quality.etl.air_quality_index.pollutant_type import PollutantType
+from air_quality.etl.forecast.forecast_data import (
     convert_east_only_longitude_to_east_west,
     ForecastData,
 )
