+++ air-quality-backend/tests/etl_tests/in_situ/sort_in_situ_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from etl.in_situ.sort_in_situ import sort_by_distance_and_time
+from air_quality.etl.in_situ.sort_in_situ import sort_by_distance_and_time
 from tests.etl_tests.in_situ.mock_openaq_data import (
     openaq_dataset_to_sort_distance,
     openaq_dataset_to_sort_time,
